+++ src/components.d.ts
@@ -16,12 +16,12 @@ import {
   RestFetch,
 } from './utils/restFetch';
 import {
+  Marketplace,
+} from './types/marketplace';
+import {
   GraphqlRequestBody,
   GraphqlResponseBody,
 } from './utils/graphqlFetch';
-import {
-  Marketplace,
-} from './types/marketplace';
 import {
   Option,
 } from './types/Select';
@@ -88,6 +88,11 @@ export namespace Components {
     * _(hidden)_ Passed by `<manifold-connection>`
     */
     'restFetch'?: RestFetch;
+  }
+  interface ManifoldCredentialsView {
+    'credentials'?: Marketplace.Credential[];
+    'loading': boolean;
+    'resourceLabel': string;
   }
   interface ManifoldDataDeprovisionButton {
     'loading'?: boolean;
@@ -497,11 +502,9 @@ export namespace Components {
     */
     'restFetch'?: RestFetch;
   }
-  interface ManifoldResourceCredentials {}
-  interface ManifoldResourceCredentialsView {
-    'credentials'?: Marketplace.Credential[];
+  interface ManifoldResourceCredentials {
+    'data'?: Gateway.Resource;
     'loading': boolean;
-    'resourceLabel': string;
   }
   interface ManifoldResourceDeprovision {
     'data'?: Gateway.Resource;
@@ -668,6 +671,12 @@ declare global {
   var HTMLManifoldCredentialsElement: {
     prototype: HTMLManifoldCredentialsElement;
     new (): HTMLManifoldCredentialsElement;
+  };
+
+  interface HTMLManifoldCredentialsViewElement extends Components.ManifoldCredentialsView, HTMLStencilElement {}
+  var HTMLManifoldCredentialsViewElement: {
+    prototype: HTMLManifoldCredentialsViewElement;
+    new (): HTMLManifoldCredentialsViewElement;
   };
 
   interface HTMLManifoldDataDeprovisionButtonElement extends Components.ManifoldDataDeprovisionButton, HTMLStencilElement {}
@@ -862,12 +871,6 @@ declare global {
     new (): HTMLManifoldResourceCredentialsElement;
   };
 
-  interface HTMLManifoldResourceCredentialsViewElement extends Components.ManifoldResourceCredentialsView, HTMLStencilElement {}
-  var HTMLManifoldResourceCredentialsViewElement: {
-    prototype: HTMLManifoldResourceCredentialsViewElement;
-    new (): HTMLManifoldResourceCredentialsViewElement;
-  };
-
   interface HTMLManifoldResourceDeprovisionElement extends Components.ManifoldResourceDeprovision, HTMLStencilElement {}
   var HTMLManifoldResourceDeprovisionElement: {
     prototype: HTMLManifoldResourceDeprovisionElement;
@@ -990,6 +993,7 @@ declare global {
     'manifold-connection': HTMLManifoldConnectionElement;
     'manifold-cost-display': HTMLManifoldCostDisplayElement;
     'manifold-credentials': HTMLManifoldCredentialsElement;
+    'manifold-credentials-view': HTMLManifoldCredentialsViewElement;
     'manifold-data-deprovision-button': HTMLManifoldDataDeprovisionButtonElement;
     'manifold-data-has-resource': HTMLManifoldDataHasResourceElement;
     'manifold-data-manage-button': HTMLManifoldDataManageButtonElement;
@@ -1022,7 +1026,6 @@ declare global {
     'manifold-resource-card-view': HTMLManifoldResourceCardViewElement;
     'manifold-resource-container': HTMLManifoldResourceContainerElement;
     'manifold-resource-credentials': HTMLManifoldResourceCredentialsElement;
-    'manifold-resource-credentials-view': HTMLManifoldResourceCredentialsViewElement;
     'manifold-resource-deprovision': HTMLManifoldResourceDeprovisionElement;
     'manifold-resource-details': HTMLManifoldResourceDetailsElement;
     'manifold-resource-details-view': HTMLManifoldResourceDetailsViewElement;
@@ -1110,6 +1113,12 @@ declare namespace LocalJSX {
     * _(hidden)_ Passed by `<manifold-connection>`
     */
     'restFetch'?: RestFetch;
+  }
+  interface ManifoldCredentialsView extends JSXBase.HTMLAttributes<HTMLManifoldCredentialsViewElement> {
+    'credentials'?: Marketplace.Credential[];
+    'loading'?: boolean;
+    'onCredentialsRequested'?: (event: CustomEvent<any>) => void;
+    'resourceLabel'?: string;
   }
   interface ManifoldDataDeprovisionButton extends JSXBase.HTMLAttributes<HTMLManifoldDataDeprovisionButtonElement> {
     'loading'?: boolean;
@@ -1543,12 +1552,9 @@ declare namespace LocalJSX {
     */
     'restFetch'?: RestFetch;
   }
-  interface ManifoldResourceCredentials extends JSXBase.HTMLAttributes<HTMLManifoldResourceCredentialsElement> {}
-  interface ManifoldResourceCredentialsView extends JSXBase.HTMLAttributes<HTMLManifoldResourceCredentialsViewElement> {
-    'credentials'?: Marketplace.Credential[];
+  interface ManifoldResourceCredentials extends JSXBase.HTMLAttributes<HTMLManifoldResourceCredentialsElement> {
+    'data'?: Gateway.Resource;
     'loading'?: boolean;
-    'onCredentialsRequested'?: (event: CustomEvent<any>) => void;
-    'resourceLabel'?: string;
   }
   interface ManifoldResourceDeprovision extends JSXBase.HTMLAttributes<HTMLManifoldResourceDeprovisionElement> {
     'data'?: Gateway.Resource;
@@ -1688,6 +1694,7 @@ declare namespace LocalJSX {
     'manifold-connection': ManifoldConnection;
     'manifold-cost-display': ManifoldCostDisplay;
     'manifold-credentials': ManifoldCredentials;
+    'manifold-credentials-view': ManifoldCredentialsView;
     'manifold-data-deprovision-button': ManifoldDataDeprovisionButton;
     'manifold-data-has-resource': ManifoldDataHasResource;
     'manifold-data-manage-button': ManifoldDataManageButton;
@@ -1720,7 +1727,6 @@ declare namespace LocalJSX {
     'manifold-resource-card-view': ManifoldResourceCardView;
     'manifold-resource-container': ManifoldResourceContainer;
     'manifold-resource-credentials': ManifoldResourceCredentials;
-    'manifold-resource-credentials-view': ManifoldResourceCredentialsView;
     'manifold-resource-deprovision': ManifoldResourceDeprovision;
     'manifold-resource-details': ManifoldResourceDetails;
     'manifold-resource-details-view': ManifoldResourceDetailsView;