+++ src/components/manifold-credentials-view/manifold-credentials-view.tsx
@@ -1,19 +1,38 @@
-import { h, Component, Prop, Event, EventEmitter } from '@stencil/core';
+import { h, Component, Prop, Event, EventEmitter, Element } from '@stencil/core';
 import { eye, lock, loader } from '@manifoldco/icons';
 
 import { Marketplace } from '../../types/marketplace';
 import logger from '../../utils/logger';
 
 @Component({
-  tag: 'manifold-resource-credentials-view',
+  tag: 'manifold-credentials-view',
   styleUrl: 'manifold-credentials-view.css',
   shadow: true,
 })
 export class ManifoldResourceCredentials {
+  @Element() private el: HTMLElement;
   @Prop() credentials?: Marketplace.Credential[];
   @Prop() resourceLabel: string = '';
   @Prop() loading: boolean = false;
   @Event() credentialsRequested: EventEmitter;
+
+  showButtonEl?: Element;
+  hideButtonEl?: Element;
+
+  componentWillLoad() {
+    this.findNodes();
+    this.addListeners();
+  }
+
+  componentDidUpdate() {
+    this.removeListeners();
+    this.findNodes();
+    this.addListeners();
+  }
+
+  componentDidUnload() {
+    this.removeListeners();
+  }
 
   get lines() {
     // Returns # of lines for creds.
@@ -27,6 +46,36 @@ export class ManifoldResourceCredentials {
     );
   }
 
+  findNodes() {
+    this.el.childNodes.forEach(child => {
+      if ((child as HTMLElement).slot === 'show-button') {
+        this.showButtonEl =
+          (child as HTMLElement).querySelector(':not(manifold-forward-slot)') || undefined;
+      } else if ((child as HTMLElement).slot === 'hide-button') {
+        this.hideButtonEl =
+          (child as HTMLElement).querySelector(':not(manifold-forward-slot)') || undefined;
+      }
+    });
+  }
+
+  addListeners() {
+    if (this.showButtonEl) {
+      this.showButtonEl.addEventListener('click', this.requestCredentials);
+    }
+    if (this.hideButtonEl) {
+      this.hideButtonEl.addEventListener('click', this.hideCredentials);
+    }
+  }
+
+  removeListeners() {
+    if (this.showButtonEl) {
+      this.showButtonEl.removeEventListener('click', this.requestCredentials);
+    }
+    if (this.hideButtonEl) {
+      this.hideButtonEl.removeEventListener('click', this.hideCredentials);
+    }
+  }
+
   hideCredentials = () => {
     this.credentials = undefined;
   };
@@ -37,12 +86,24 @@ export class ManifoldResourceCredentials {
 
   @logger()
   render() {
+    const showButton = this.showButtonEl ? (
+      <slot name="show-button" />
+    ) : (
+      <manifold-button color="black" stencilClickEvent={this.requestCredentials}>
+        <manifold-icon marginRight icon={eye} /> Show credentials
+      </manifold-button>
+    );
+
     return [
       <menu class="secrets-menu" data-showing={!!this.credentials}>
-        <manifold-button color="white" size="small" stencilClickEvent={this.hideCredentials}>
-          <manifold-icon marginRight icon={lock} />
-          Hide credentials
-        </manifold-button>
+        {this.hideButtonEl ? (
+          <slot name="hide-button" />
+        ) : (
+          <manifold-button color="white" size="small" stencilClickEvent={this.hideCredentials}>
+            <manifold-icon marginRight icon={lock} />
+            Hide credentials
+          </manifold-button>
+        )}
       </menu>,
       <div
         class="credential"
@@ -78,9 +139,7 @@ export class ManifoldResourceCredentials {
               </span>
             </manifold-button>
           ) : (
-            <manifold-button color="black" stencilClickEvent={this.requestCredentials}>
-              <manifold-icon marginRight icon={eye} /> Show credentials
-            </manifold-button>
+            showButton
           )}
         </div>
       </div>,